app/requirements.txt
# ============================================================================
# vLLM Docker 클라이언트 의존성
# ============================================================================
# vLLM 서버는 Docker로 실행되므로 클라이언트는 경량 패키지만 설치

# UI Framework
gradio==4.44.0

# OpenAI API 클라이언트 (vLLM 서버와 통신)
openai>=1.3.0,<2.0.0

# Configuration
python-dotenv==1.0.1

# Utilities
requests==2.32.3
